Started work on Twitter integration for phone and address lookup.

+++ config/initializers/_settings.rb
@@ -1,8 +1,10 @@
 if RAILS_ENV == 'test'
   %w(settings sites).each do |file|
+    model = eval(file.singularize.classify)
+    next unless model.table_exists?
     YAML::load(File.open(File.join(RAILS_ROOT, "test/fixtures/#{file}.yml"))).each do |fixture, values|
-      eval(file.singularize.classify).create(values)
+      model.create(values)
     end
   end
-  Site.current = Site.find(1)
+  Site.current = Site.find(1) if Site.table_exists?
 end
